+++ OpERP/src/main/java/devopsdistilled/operp/client/items/models/CategoryModel.java
@@ -1,7 +1,7 @@
 package devopsdistilled.operp.client.items.models;
 
 import devopsdistilled.operp.client.abstracts.EntityModel;
-import devopsdistilled.operp.client.items.panes.models.observers.CategoryModelObserver;
+import devopsdistilled.operp.client.items.models.observers.CategoryModelObserver;
 import devopsdistilled.operp.server.data.entity.items.Category;
 import devopsdistilled.operp.server.data.service.items.CategoryService;
