+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/ListItemPane.java
@@ -12,8 +12,8 @@ import net.miginfocom.swing.MigLayout;
 import devopsdistilled.operp.client.abstracts.SubTaskPane;
 import devopsdistilled.operp.client.items.models.impl.ItemTableModel;
 import devopsdistilled.operp.client.items.models.observers.ItemModelObserver;
-import devopsdistilled.operp.client.items.models.observers.ListItemPaneModelObserver;
 import devopsdistilled.operp.client.items.panes.controllers.ListItemPaneController;
+import devopsdistilled.operp.client.items.panes.models.observers.ListItemPaneModelObserver;
 import devopsdistilled.operp.server.data.entity.items.Item;
 
 public class ListItemPane extends SubTaskPane implements
