+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/models/CreateItemPaneModel.java
@@ -1,7 +1,7 @@
 package devopsdistilled.operp.client.items.panes.models;
 
 import devopsdistilled.operp.client.abstracts.PaneModel;
-import devopsdistilled.operp.client.items.models.observers.CreateItemPaneModelObserver;
+import devopsdistilled.operp.client.items.panes.models.observers.CreateItemPaneModelObserver;
 
 public interface CreateItemPaneModel extends
 		PaneModel<CreateItemPaneModelObserver> {
