+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/models/EditItemPaneModel.java
@@ -1,7 +1,7 @@
 package devopsdistilled.operp.client.items.panes.models;
 
 import devopsdistilled.operp.client.abstracts.PaneModel;
-import devopsdistilled.operp.client.items.models.observers.EditItemPaneModelObserver;
+import devopsdistilled.operp.client.items.panes.models.observers.EditItemPaneModelObserver;
 import devopsdistilled.operp.server.data.entity.items.Item;
 
 public interface EditItemPaneModel extends PaneModel<EditItemPaneModelObserver> {
