+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/models/ListItemPaneModel.java
@@ -1,7 +1,7 @@
 package devopsdistilled.operp.client.items.panes.models;
 
 import devopsdistilled.operp.client.abstracts.PaneModel;
-import devopsdistilled.operp.client.items.models.observers.ListItemPaneModelObserver;
+import devopsdistilled.operp.client.items.panes.models.observers.ListItemPaneModelObserver;
 
 public interface ListItemPaneModel extends PaneModel<ListItemPaneModelObserver> {
