+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/models/observers/ListItemPaneModelObserver.java
@@ -1,4 +1,4 @@
-package devopsdistilled.operp.client.items.models.observers;
+package devopsdistilled.operp.client.items.panes.models.observers;
 
 import devopsdistilled.operp.client.abstracts.PaneModelObserver;
